Application_practice/requirements.txt
easyocr==1.7.1
groq==0.11.0
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
Pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
scikit_learn==1.5.2
seaborn==0.13.2
streamlit==1.37.1
streamlit_extras==0.4.7
